requirements.txt
click>=8.1
rich>=13.0
anthropic>=0.40
openai>=1.0
pyyaml>=6.0
